combine.twosamples {twosamples}	R Documentation

Combine two objects of class twosamples

Description

This function combines two twosamples objects – concatenating bootstraps, recalculating pvalues, etc. It only works if both objects were created with "keep.boots=T" This function is intended for one main purposes: combining parallized null calculations and then plotting those combined outputs.

Usage

combine.twosamples(x, y, check.sample = T)

Arguments

x
a twosamples object

y
a different twosamples object from the same *_test function run on the same data

check.sample
check that the samples saved in each object are the same? (can be slow)

Value

a twosamples object that correctly re-calculates the p-value and determines all the other attributes

See Also

twosamples_class, plot.twosamples, dts_test

Examples

vec1 = rnorm(10)
vec2 = rnorm(10,1)
out1 = dts_test(vec1,vec2)
out2 = dts_test(vec1,vec2)
combined = combine.twosamples(out1,out2)
summary(out1)
summary(out2)
summary(combined)
plot(combined)

[Package twosamples version 2.0.1 Index]