plot.twosamples {twosamples}	R Documentation

Default plots for twosamples objects

Description

Typically for now this will produce a histogram of the null distribution based on the bootstrapped values, with a vertical line marking the value of the test statistic.

Usage

## S3 method for class 'twosamples'
plot(x, plot_type = c("boots_hist"), nbins = 50, ...)

Arguments

x
an object produced by one of the twosamples *_test functions

plot_type
which plot to create? only current option is "boots_hist",

nbins
how many bins (or breaks) in the histogram

...
other parameters to be passed to plotting functions

Value

Produces a plot

See Also

dts_test(), twosamples_class, combine.twosamples

Examples

out = dts_test(rnorm(10),rnorm(10,1))
plot(out)

[Package twosamples version 2.0.1 Index]