print.twosamples {twosamples}	R Documentation

twosamples_class

Description

Objects of Class twosamples are output by all of the *_test functions in the twosamples package.

Usage

## S3 method for class 'twosamples'
print(x, ...)

## S3 method for class 'twosamples'
summary(object, alpha = 0.05, ...)

Arguments

x
twosamples object

...
other parameters to be passed to print or summary functions

object
twosamples-object to summarize

alpha
Significance threshold for determining null rejection

Details

By default they consist of:a length 2 vector, the first item being the test statistic, the second the p-value. That vector has the following attributes:

details: length 3 vector with the sample sizes for each sample and the number of bootstraps

test_type: a string describing the type of the test statistic

It may also have two more attributes, depending on options used when running the *_test function. These are useful for plotting and combining test runs.

bootstraps: a vector containing all the bootstrapped null values

samples: a list containing both the samples that were tested

and by virtue of being a named length 2 vector of class "twosamples" it has the following two attributes:

names: c("Test Stat","P-Value")

class: "twosamples"

Multiple Twosamples objects made by the same *_test routine being run on the same data can be combined (getting correct p-value and correct attributes) with the function combine_twosamples().

Value

print.twosamples() returns nothing

summarize.twosamples() returns nothing

Functions

print(twosamples): Print method for objects of class twosamples

summary(twosamples): Summary method for objects of class twosamples

See Also

plot.twosamples(), combine.twosamples()

[Package twosamples version 2.0.1 Index]