Permutation Test Builder

This function takes a simple two-sample test statistic and produces a function which performs randomization tests (sampling with replacement) using that test stat.

Usage

permutation_test_builder(test_stat_function, default.p = 2)

Arguments

test_stat_function

a function of two vectors producing a positive number, intended as the test-statistic to be used.

default.p

This allows for some introduction of defaults and parameters. Typically used to control the power functions raise something to.

Value

This function returns a function which will perform permutation tests on given test stat.

Details

test_stat_function must be structured to take two separate vectors, and then a third value. i.e. (fun = function(vec1,vec2,val1) ...). See examples.

Functions

permutation_test_builder: Takes a test statistic, returns a testing function.

See also

Examples

mean_stat = function(a,b,p) abs(mean(a)-mean(b))**p
myfun = permutation_test_builder(mean_stat,2.0)
vec1 = rnorm(20)
vec2 = rnorm(20,4)
myfun(vec1,vec2)
#> Test Stat   P-Value 
#>  16.54753   0.00025 
#> attr(,"details")
#>      n1      n2 n.boots 
#>      20      20    2000